split out anon user manager for easier testing
