+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/WebEndpoint.kt
@@ -6,7 +6,7 @@ import org.vaccineimpact.orderlyweb.models.PermissionRequirement
 import org.vaccineimpact.orderlyweb.security.OrderlyWebSecurityLogic
 import org.vaccineimpact.orderlyweb.security.SkipOptionsMatcher
 import org.vaccineimpact.orderlyweb.security.WebSecurityConfigFactory
-import org.vaccineimpact.orderlyweb.security.authentication.AuthenticationConfig
+import org.vaccineimpact.orderlyweb.security.authentication.OrderlyWebAuthenticationConfig
 import org.vaccineimpact.orderlyweb.security.authentication.AuthenticationProvider
 import org.vaccineimpact.orderlyweb.security.clients.OrderlyWebIndirectClient
 import spark.route.HttpMethod
@@ -24,7 +24,7 @@ data class WebEndpoint(
         val externalAuth: Boolean = false,
         val spark: SparkWrapper = SparkServiceWrapper(),
         val configFactory: ConfigFactory? = null,
-        val authenticationConfig: AuthenticationConfig = AuthenticationConfig()
+        val authenticationConfig: OrderlyWebAuthenticationConfig = OrderlyWebAuthenticationConfig()
 ) : EndpointDefinition
 {
     override val allowParameterAuthentication = false
@@ -40,8 +40,8 @@ data class WebEndpoint(
 
     private fun addSecurityFilter(url: String)
     {
-        //If Montagu Auth, OrderlyWeb auth should be fully synchronised with the external login provider, and login
-        //page should not be seen
+        //If Montagu Auth and if anon users are not allowed, OrderlyWeb auth should be fully synchronised
+        // with the external login provider, and login page should not be seen
         val synchronisedAuth = authenticationConfig.getConfiguredProvider() == AuthenticationProvider.Montagu
                 && !authenticationConfig.allowAnonUser
 
@@ -65,11 +65,7 @@ data class WebEndpoint(
                 client.javaClass.simpleName,
                 config.authorizers.map { it.key }.joinToString(","),
                 SkipOptionsMatcher.name
-        )
-        if (authenticationConfig.allowAnonUser)
-        {
-            filter.apply { securityLogic = OrderlyWebSecurityLogic() }
-        }
+        ).apply { securityLogic = OrderlyWebSecurityLogic() }
         spark.before(url, filter)
     }
 }
