+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/app_start/AuthenticationRouteBuilder.kt
@@ -3,7 +3,7 @@ package org.vaccineimpact.orderlyweb.app_start
 import org.pac4j.sparkjava.CallbackRoute
 import org.pac4j.sparkjava.LogoutRoute
 import org.vaccineimpact.orderlyweb.security.WebSecurityConfigFactory
-import org.vaccineimpact.orderlyweb.security.authentication.AuthenticationConfig
+import org.vaccineimpact.orderlyweb.security.authentication.OrderlyWebAuthenticationConfig
 import org.vaccineimpact.orderlyweb.security.authentication.AuthenticationProvider
 
 interface AuthenticationRouteBuilder
@@ -12,7 +12,7 @@ interface AuthenticationRouteBuilder
     fun loginCallback(): CallbackRoute
 }
 
-class OrderlyAuthenticationRouteBuilder(private val authenticationConfig: AuthenticationConfig)
+class OrderlyAuthenticationRouteBuilder(private val authenticationConfig: OrderlyWebAuthenticationConfig)
     : AuthenticationRouteBuilder
 {
     private val client = authenticationConfig.getAuthenticationIndirectClient()
